+++ lib/lorekeeper/fast_logger.rb
@@ -69,8 +69,7 @@ module Lorekeeper
     # This is part of the standard Logger API, we need this to be compatible
     def add(severity, message_param = nil, progname = nil, &block)
       return true if severity < @level
-
-      message = message_param || block&.call || progname
+      message = (block && block.call) || message_param || progname
       log_data(severity, message.freeze)
     end
 
@@ -89,6 +88,10 @@ module Lorekeeper
 
     # inherited classes probably want to reimplement this
     def log_data(_severity, message)
+      write(message)
+    end
+
+    def write(message)
       @iodevice.write(message)
     end