+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 2.3.0
+* Add `#write` method support to all loggers
+
+# 2.2.1
+* Fix MultiLogger to not raise NoMethodError
+
+# 2.2.0
+* Remove metaprogramming from MultiLogger
+* Drop support for Ruby < 2.7.0
+
 # 2.1.0
 * Modify SimpleLogger to properly log exceptions with named parameters